common_struct 헤더파일 분리 작업
- diff_check common_utility 선언 및 정의 부분 제거
- 불 필요한 헤더 include 삭제
