+++ item_dao_sqlite.h
@@ -9,7 +9,7 @@
 #include "sqlite_wrapper.h"
 
 #include "item_dao.h"
-#include "common_utility.h"
+#include "common_struct.h"
 
 namespace monitor_client {
 	class ItemDaoSqlite : public ItemDao {
